# docs/swagger.yaml
basePath: /api
definitions:
  ad.CreateRequest:
    properties:
      description:
        type: string
      price:
        type: number
      title:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact: {}
  description: API Server for Marketplace for influences and companies
  title: Marketplace
  version: "1.0"
paths:
  /ad/{id}:
    get:
      consumes:
      - application/json
      description: Get ad details by its ID
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get ad by ID
      tags:
      - ad
  /ad/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an ad (e.g., open/closed)
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: string
      - description: Ad Status
        in: body
        name: status
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update ad status
      tags:
      - ad
  /ad/all:
    get:
      consumes:
      - application/json
      description: Get a list of all ads
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all ads
      tags:
      - ad
  /ad/company/{id}:
    get:
      consumes:
      - application/json
      description: Get all ads for a specific company
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get ads by company ID
      tags:
      - ad
  /ad/create:
    post:
      consumes:
      - application/json
      description: Create a new ad with the input payload
      parameters:
      - description: Ad Creation Request
        in: body
        name: ad
        required: true
        schema:
          $ref: '#/definitions/ad.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new ad
      tags:
      - ad
  /ad/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an ad by its ID
      parameters:
      - description: Ad ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete an ad
      tags:
      - ad
...
securityDefinitions:
  BearerAuth:
    description: JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"